Added Texture and Collision classes, almost finished TextureManager static class for Carbon support.

+++ Nikki/Reflection/Enum/eTextureMipmapBiasType.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Enum of possible mipmap bias types.
     /// </summary>
-    public enum eTextureMipmapBiasType : int
+    public enum eTextureMipmapBiasType : byte
     {
         /// <summary>
         ///
